backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
boto3==1.34.0  # AWS SDK for DynamoDB
# psycopg2-binary==2.9.9  # For Redshift if needed - commented out due to build issues

# Data processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Task queue
celery==5.3.4
redis==5.0.1

# Cloud provider SDKs
azure-identity==1.15.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-resource==23.0.1
google-cloud-billing==1.11.2
google-cloud-compute==1.14.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
